moved offcavas from components to layout
